requirements.txt
tensorflow==2.15.0
mediapipe==0.10.10
opencv-python==4.8.0.74

numpy==1.26.4
scipy==1.11.4

pandas==2.1.4
scikit-learn==1.3.2

matplotlib==3.8.2
seaborn==0.13.1
Pillow==10.2.0
tqdm==4.66.1
pyyaml==6.0.1

pytest==7.4.4
python-dotenv==1.0.0
